# src/interfaces/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Gravity PMS API
  description: Enterprise Project Management System — REST API v1
  version: 1.0.0
  contact:
    name: Gravity Platform
    email: admin@gravity.io
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.gravity.io/api/v1
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks with filters
      parameters:
        - in: query
          name: projectId
          schema: { type: string, format: uuid }
        - in: query
          name: sprintId
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { type: string, enum: [todo, in_progress, done] }
        - in: query
          name: priority
          schema: { type: string, enum: [low, medium, high, critical] }
      responses:
        '200':
          description: Task list
    post:
      tags: [Tasks]
      summary: Create task
      responses:
        '201':
          description: Created task
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task with comments, attachments, history
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Task detail
    put:
      tags: [Tasks]
      summary: Update task (status, assignee, etc.)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Updated
    delete:
      tags: [Tasks]
      summary: Delete task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Deleted
  /tasks/{id}/comments:
    post:
      tags: [Tasks]
      summary: Add comment to task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                is_qa: { type: boolean, default: false }
      responses:
        '201':
          description: Comment created
  /tasks/{id}/attachments:
    post:
      tags: [Tasks]
      summary: Upload attachment to task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Attachment uploaded
...